refactor: move getStaff of each superviosr and clevel to UserController
