+++ app/Http/Controllers/ReportController.php
@@ -13,16 +13,6 @@ use App\Http\Resources\DailyReportResource;
 
 class ReportController extends Controller
 {
-    public function getStaffOfSupervisor()
-    {
-        $supervisorId = Auth::id();
-        $staff = User::where('supervisor_id', $supervisorId)
-            ->select('id', 'profile_picture', 'first_name', 'last_name')
-            ->get();
-
-        return UserResource::collection($staff);
-    }
-
     public function getWeeklyReport(Request $request)
     {
         // Logic to fetch and filter weekly report
@@ -87,29 +77,4 @@ class ReportController extends Controller
             'weekly_report_completion_percentage' => round($completionPercentage, 2)
         ]);
     }
-    
-    public function getCLevelStaff()
-    {
-        $cLevelId = Auth::id();
-
-        // Get supervisors under the C-Level
-        $supervisors = User::where('supervisor_id', $cLevelId)
-            ->select('id', 'profile_picture', 'first_name', 'last_name')
-            ->with(['staff' => function($query) {
-                $query->select('id', 'profile_picture', 'first_name', 'last_name', 'supervisor_id');
-            }])
-            ->get();
-
-        return response()->json([
-            'c_level_id' => $cLevelId,
-            'supervisors' => $supervisors->map(function ($supervisor) {
-                return [
-                    'supervisor' => new UserResource($supervisor),
-                    'staff' => UserResource::collection($supervisor->staff)
-                    
-                ];
-            })
-        ]);
-    }
-
 }
